🧑‍💻 (storybook): Add storybook stuff to build
