+++ src/forms/FGenericForm/sb.stuff.js
@@ -1,4 +1,9 @@
+import { computed, ref } from 'vue';
+import { useArgs } from 'storybook/preview-api';
+
 import { LAYOUT_CHOICES } from '@/forms/FFormRow/constants.js';
+import { makeUpdateArg } from '@/sb.stuff.js';
+import FGenericForm from '.';
 import { INPUT_WIDGET_TYPES, SIZE_CHOICES } from './constants.js';
 
 
@@ -126,3 +131,80 @@ export const META_BASIC = [
 		fields: [{ fieldName: 'user_type', default: 'user', required: true }],
 	},
 ];
+
+
+/**
+ * Render function for single-widget stories inside FGenericForm.
+ *
+ * Collects all args whose argType has `table.category === 'meta'` and assembles them
+ * into the single meta entry object, then wraps it in an array for FGenericForm.
+ * Forwards `layout`, `size → widgetSize`, `fieldErrors → errors`, and `widgets`.
+ * The `modelValue` arg is kept in sync with Storybook controls via `makeUpdateArg`.
+ *
+ * Use for Default / Errors stories where a single widget instance is shown
+ * and the user can interact with the form via Storybook controls.
+ */
+export const makeFGFWidgetRenderer = () => {
+	return (args, { argTypes }) => {
+		const [ , updateArgs ] = useArgs();
+		const metaKeys = Object.keys(argTypes).filter((k) => {
+			const categoryOk = argTypes[k]?.table?.category === 'props';
+			const subcategoryOk = argTypes[k]?.table?.subcategory === 'meta';
+			return categoryOk && subcategoryOk;
+		});
+		return {
+			components: { FGenericForm },
+			setup() {
+				const modelValueArg = makeUpdateArg('modelValue', updateArgs);
+				const formArgs = computed(() => {
+					const meta = Object.fromEntries(metaKeys.map(k => [k, args[k]]));
+					return {
+						'modelValue': args.modelValue,
+						[modelValueArg[0]]: modelValueArg[2],
+						'meta': [meta],
+						'layout': args.layout,
+						'widgetSize': args.size,
+						'errors': args.fieldErrors,
+						'widgets': args.widgets,
+					};
+				});
+
+				return { formArgs };
+			},
+			template: '<FGenericForm v-bind="formArgs" />',
+		};
+	};
+};
+
+
+/**
+ * Render function for multi-widget stories inside FGenericForm.
+ *
+ * Expects `meta` to already be an array of entries and mounts a single
+ * `FGenericForm` with all of them. Forwards `layout`, `size → widgetSize`,
+ * `fieldErrors → errors`, and `widgets`.
+ *
+ * Unlike `makeFGFWidgetRenderer`, `modelValue` is held in a local `ref` and is
+ * NOT synced back to Storybook controls — intentional for static demo stories
+ * (States, Types) where field controls are disabled and the initial values are
+ * all that matter.
+ *
+ * Use when showing multiple field variants side by side with fixed initial values.
+ */
+export const makeFGFWidgetManyRenderer = () => {
+	return (args) => ({
+		components: { FGenericForm },
+		setup() {
+			const modelValue = ref(args.modelValue);
+			const formArgs = computed(() => ({
+				'meta': args.meta,
+				'layout': args.layout,
+				'widgetSize': args.size,
+				'errors': args.fieldErrors,
+				'widgets': args.widgets,
+			}));
+			return { modelValue, formArgs };
+		},
+		template: '<FGenericForm v-model="modelValue" v-bind="formArgs" />',
+	});
+};
