+++ src/forms/FGenericForm/widgets/FRadioButtonWidget.stories.js
@@ -1,5 +1,4 @@
-import { makeFGFWidgetRenderer, makeFGFWidgetManyRenderer } from '@/.storybook/utils.js';
-import { WIDGET_BASE_ARG_TYPES } from '../constants.sb.js';
+import { WIDGET_BASE_ARG_TYPES, makeFGFWidgetRenderer, makeFGFWidgetManyRenderer } from '../sb.stuff.js';
 import FRadioButtonWidgetComponent from './FRadioButtonWidget.vue';
 import { getFormDefaults } from '../utils';
 import { DEFAULT_WIDGETS } from '../constants.js';
