+++ src/forms/FGenericForm/widgets/FSelectWidget.stories.js
@@ -1,5 +1,4 @@
-import { makeFGFWidgetRenderer, makeFGFWidgetManyRenderer } from '@/.storybook/utils.js';
-import { WIDGET_BASE_ARG_TYPES } from '../constants.sb.js';
+import { WIDGET_BASE_ARG_TYPES, makeFGFWidgetRenderer, makeFGFWidgetManyRenderer } from '../sb.stuff.js';
 import { DEFAULT_WIDGETS } from '../constants.js';
 import FSelectWidgetComponent from './FSelectWidget.vue';
 import { getFormDefaults } from '../utils';
