+++ src/forms/FGenericForm/widgets/FTextareaWidget.stories.js
@@ -1,5 +1,4 @@
-import { makeFGFWidgetRenderer, makeFGFWidgetManyRenderer } from '@/.storybook/utils.js';
-import { WIDGET_BASE_ARG_TYPES } from '../constants.sb.js';
+import { WIDGET_BASE_ARG_TYPES, makeFGFWidgetRenderer, makeFGFWidgetManyRenderer } from '../sb.stuff.js';
 import { DEFAULT_WIDGETS } from '../constants.js';
 import FTextareaWidgetComponent from './FTextareaWidget.vue';
 import { getFormDefaults } from '../utils';
