+++ src/forms/FTextarea/FTextarea.stories.js
@@ -1,7 +1,7 @@
 import { computed, ref } from 'vue';
 import { useArgs } from 'storybook/preview-api';
 
-import { makeUpdateArg } from '@/.storybook/utils.js';
+import { makeUpdateArg } from '@/sb.stuff.js';
 import FTextarea from './index.vue';
 import { SIZE_CHOICES } from './constants';
 import Readme from './README.md?raw';
