+++ src/sb.stuff.js
@@ -1,0 +1,78 @@
+import { computed } from 'vue';
+import { action } from 'storybook/actions';
+import { useArgs } from 'storybook/preview-api';
+import throttle from 'lodash/throttle';
+
+
+/**
+ * Creates a handler to synchronize v-model changes in the component
+ * with Storybook's args state.
+ *
+ * Consider making the throttle configurable if you need different update frequencies.
+ *
+ * @param {string | [string, string] | [string, string, string]} argOption - Either the name of the argument to update,
+ *   or an array with the argument name, argument for listener name and the event name (e.g., `['modelValue', 'onUpdate:modelValue', 'update:modelValue']`).
+ * @param {Object} args - All story args from the render function
+ *   (`render: (args, { argTypes, updateArgs }) => { ... }`).
+ * @param {Function} updateArgs - Storybook's function for updating args
+ *   (`render: (args, { argTypes, updateArgs }) => { ... }`).
+ *
+ * @returns {[string, string, Function]} An array with:
+ *   - the arg name for listener emits value changes
+ *   - the event name that emits value changes
+ *   - a throttled function that updates the Storybook arg
+ */
+export const makeUpdateArg = (argOption, updateArgs) => {
+	const arg = Array.isArray(argOption) ? argOption[0] : argOption;
+	const updateArg = Array.isArray(argOption) ? argOption[1] : `onUpdate:${arg}`;
+	const updateEvent = Array.isArray(argOption) ? argOption[2] : `update:${arg}`;
+	return [
+		updateArg,
+		updateEvent,
+		(newValue) => {
+			action(updateArg)(newValue);
+			if (updateArgs) {
+				setTimeout(() => { updateArgs({ [arg]: newValue }) }, 0);
+			}
+		},
+	];
+};
+
+
+/**
+ * Helper to generate a render function for a Storybook story
+ * that synchronizes the specified list of arguments (see `makeUpdateArg`).
+ *
+ * @param {Array<string | [string, string] | [string, string, string]>} updatableArgs - List of arguments to synchronize.
+ *
+ * @returns {Function} A render function for the Storybook story.
+ */
+export const makeRenderer = (updatableArgs) => {
+	return (args, { argTypes, component }) => {
+		const [ , updateArgs ] = useArgs();
+		return {
+			components: { [component.name]: component },
+			props: Object.keys(argTypes),
+			setup() {
+				const tmpUpdatableArgs = updatableArgs.map(item => {
+					return makeUpdateArg(item, updateArgs);
+				});
+
+				const newArgs = computed(() => {
+					const result = { ...args };
+
+					tmpUpdatableArgs.forEach(([ argKey, eventKey, value ]) => {
+						delete result[eventKey];
+						result[argKey] = value;
+					});
+
+					return result;
+				});
+
+
+				return { args: newArgs };
+			},
+			template: `<${component.name} v-bind="args" />`,
+		};
+	};
+};
